+++ openpeer/services/ISTUNRequesterManager.h
@@ -58,7 +58,7 @@ namespace openpeer
       static ISTUNRequesterPtr handlePacket(
                                             IPAddress fromIPAddress,
                                             const BYTE *packet,
-                                            ULONG packetLengthInBytes,
+                                            size_t packetLengthInBytes,
                                             STUNPacket::RFCs allowedRFCs
                                             );
