+++ openpeer/services/internal/services_RUDPICESocketSession.h
@@ -141,7 +141,7 @@ namespace openpeer
         virtual void handleICESocketSessionReceivedPacket(
                                                           IICESocketSessionPtr session,
                                                           const BYTE *buffer,
-                                                          ULONG bufferLengthInBytes
+                                                          size_t bufferLengthInBytes
                                                           );
 
         virtual bool handleICESocketSessionReceivedSTUNPacket(
@@ -167,7 +167,7 @@ namespace openpeer
                                                  RUDPChannelPtr channel,
                                                  const IPAddress &remoteIP,
                                                  const BYTE *packet,
-                                                 ULONG packetLengthInBytes
+                                                 size_t packetLengthInBytes
                                                  );
 
       protected:
